DRY: Removed C++ version of llama_sampler_init_dry

+++ include/llama.h
@@ -1146,7 +1146,7 @@ extern "C" {
                              int32_t    dry_allowed_length,
                              int32_t    dry_penalty_last_n,
                           const char ** seq_breakers,
-                                 int    num_breakers);
+                              size_t    num_breakers);
 
     LLAMA_API struct llama_sampler * llama_sampler_init_logit_bias(
                              int32_t   n_vocab,
